_hp/tests/framing.sub.html

<!DOCTYPE html>
<html lang="en">
<head>
  <meta charset="UTF-8">
  <meta http-equiv="X-UA-Compatible" content="IE=edge">
  <meta name="viewport" content="width=device-width, initial-scale=1.0">
  <title>Framing Tests</title>
  <script src="/_hp/ressources/testharness.js"></script>
  <script src="/_hp/ressources/testharnessreport.js"></script>
  <script src="/_hp/ressources/store_results.sub.js"></script>
</head>

<body>

</body>


<script>
function xfo_test({ url, response, message }) {
    console.log(url, response, message);
    async_test(t => {
      const i = document.createElement("iframe");
      i.src = url + response;
  
      //waitForMessageFrom(i, t).then(t.unreached_func("Frame should not have sent a message."));
      i.onload = t.step_func_done(() => {
        //assert_equals(i.contentDocument, null);
        t.report_outcome({"frames.length": i.contentWindow.frames.length});
        assert_equals(null, null);

      });
      //waitForMessageFrom(i, t).then(t.step_func_done(e => {
      //  assert_equals(e.data, "Loaded");
      //}));
  
      document.body.append(i);
      t.add_cleanup(() => i.remove());
    }, message);
  }
  
  function waitForMessageFrom(frame, test) {
    return new Promise(resolve => {
      window.addEventListener("message", test.step_func(e => {
        console.log(e);
        if (e.source == frame.contentWindow) {
          resolve(e);
        }
      }));
    });
  }
  
  var test_declarations = [xfo_test];


for (var test of test_declarations) {
	const url = document.location.origin + '/_hp/server/responses.py?feature_group=framing&resp_id=';
	const suburl = `http://{{domains[sub]}}:{{ports[http][0]}}/_hp/server/responses.py?feature_group=framing&resp_id=`;
	const randomurl = `http://{{hosts[alt][]}}:{{ports[http][0]}}/_hp/server/responses.py?feature_group=framing&resp_id=`;
	// TODO: decide which responses to run on, ...
	var response = 1;
	for (var u of [url, suburl, randomurl]) {
		test({url: u, response: response, message: `XFO-${u}-${response}`});
	}
}


</script>
</html>
